not compiling with the static linking yet
As a package the ability to program the meson build system to produce
static links of the game footprint enables a consolidated distribution
approach for tools such as skia where the majority of the code may not
be used completely. It is a slim model. Providing effective opperation
on gentoo offers the flexibility of rubber stamp processing to add
features that provide progressive boot patterns for usage and user.

Provide that ability to organize software operations within the
environment for better committment of work, publishing and processing.
Enabing production of media at every aspect to the generalized streams
of network such as common programs of browsers offer. As well offer
other adaptable network protocols that summarize output to the rendering
library.

The ability to reuse compiled kernel code and have storage databases for
cpu models to stamp images together for multiple types of devices is the
dream of embedded and laptop, desktop services. By providing this as a
server package along with change / diff upgrade image from server, a
likeable internet and offline computer may be found. I would spot a
concern is the chromebooks ability to support more offline only
programs.

Program that detects the parameters of the kernel build, can be based on
a few requiresments such as cpu type, and hardware device scan. The rest
of the formal kernel paraemters may be derrived from software that will
be used at the application level. By providing the suite, plus games
that utilize the technology is a rationalization that may be viewed as
most games can operate within a specific api tollerance such as gl, and
drivers are present for the video card. A steam machine from gentoo
would be effective..

Wireless communication should be built in as a service provided by the
machine for online content. External storage over high speed for usb and
thunderbolt can provide a flexible necessity of cost for level of game
hardware. by controlling the machine and architecture with gen two there
are many new possibilities which can increase battery life and provide
gaming processing especially in floating point math. Scene graph
traversal, loading, and mitmapping textures. ZThis leads to an
additional graphics buffer and port commands for the video card. I would
integrate the memory access into the cpu as a memory location associated
with an area of memory, which a start names properties such as as size.